+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/container/input-wrappers/base/base.tsx
@@ -16,7 +16,7 @@ import { Subtract } from '../../../../typescript'
 
 import * as React from 'react'
 import MarionetteRegionContainer from '../../marionette-region-container'
-import withListenTo, { WithBackboneProps } from '../../backbone-container'
+import withListenTo, { WithBackboneProps } from '../../../backbone-container'
 import { hot } from 'react-hot-loader'
 const PropertyView = require('../../../../component/property/property.view.js')
 const PropertyModel = require('../../../../component/property/property.js')
