+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/container/time-settings/index.tsx
@@ -16,7 +16,7 @@ import * as React from 'react'
 import { hot } from 'react-hot-loader'
 import * as moment from 'moment'
 
-import withListenTo, { WithBackboneProps } from '../backbone-container'
+import withListenTo, { WithBackboneProps } from '../../backbone-container'
 
 import View from '../../time-settings'
