+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/metacard-interactions/index.tsx
@@ -13,7 +13,7 @@
  *
  **/
 import * as React from 'react'
-import withListenTo, { WithBackboneProps } from '../container/backbone-container'
+import withListenTo, { WithBackboneProps } from '../backbone-container'
 import { hot } from 'react-hot-loader'
 const user = require('../../component/singletons/user-instance')
