+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/multi-select-actions/container.tsx
@@ -15,7 +15,7 @@
 import * as React from 'react'
 import { hot } from 'react-hot-loader'
 import MultiSelectActionsPresentation from './presentation'
-import withListenTo, { WithBackboneProps } from '../container/backbone-container'
+import withListenTo, { WithBackboneProps } from '../backbone-container'
 
 type Props = {
   selectionInterface: any
