+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/workspace-item/container.tsx
@@ -13,7 +13,7 @@
  *
  **/
 import * as React from 'react'
-import withListenTo, { WithBackboneProps } from '../container/backbone-container'
+import withListenTo, { WithBackboneProps } from '../backbone-container'
 import WorkspacesItem from './presentation'
 const moment = require('moment')
 const wreqr = require('../../js/wreqr.js')
